requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
django-celery-results>=2.5.1
django-celery-beat>=2.6.0

# Гео-стек
geopandas>=0.14.0
osmnx>=1.7.0
gpxpy>=1.6.0
folium>=0.15.0
shapely>=2.0.0
pyproj>=3.6.0
rasterio>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
networkx>=3.0


